general visit-Method added to ITreeVisitor. AbstractTreeWalker implements ITreeVisitor, so the visit()-Method is implemented generally.
New optimizing Classes:
    - TreeWalkerSelectionPush.java
    - TreeWalkerSelectionSplit.java
    - TreeWalkerSelectionUnion.java
Optimizer updated to interact with these Classes

Two Methods added to CompilePlan to find the Parent-relations between Operators in this CompilePlan. You can delete them if these relations are added earlier.

Optimizer Test are generated:
   - TestOptimizeSQL.java

Because of interworking with the Compiler, some Functions are Dummys to be implemented as soon as the Compiler has some further functionality.